Pulled grid out into its own file since we will have more spatial APIs
